Configuracion para correr pruebas, eliminacion de edad de persona
