+++ pom.xml
@@ -72,7 +72,7 @@
                         <!-- Default configuration for running with: mvn clean javafx:run -->
                         <id>default-cli</id>
                         <configuration>
-                            <mainClass>uv.fei.tutorias.tutoriasapp/main.HelloApplication
+                            <mainClass>uv.fei.tutorias/main.HelloApplication
                             </mainClass>
                             <launcher>app</launcher>
                             <jlinkZipName>app</jlinkZipName>
