--- src/main/java/test/PersonaDAOTest.java
@@ -1,7 +1,0 @@
-package test;
-
-import static org.junit.jupiter.api.Assertions.*;
-
-class PersonaDAOTest {
-
-}
