+++ src/main/java/uv/fei/tutorias/test/PersonaDAOTest.java
@@ -1,0 +1,12 @@
+package uv.fei.tutorias.test;
+
+import org.junit.Test;
+
+class PersonaDAOTest {
+    @Test
+    public void probarText(){
+        System.out.println("Hola");
+    }
+
+
+}
